+++ internal/services/message/domain/message_duplicate_cassandra_integration_test.go
@@ -1,4 +1,4 @@
-package service
+package messagedomain
 
 import (
 	"os"